requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
cov-core==1.15.0
coverage==3.7.1
Django==3.0.6
django-jquery==3.1.0
django-phone-login==1.1.3
django-phonenumber-field==2.1.0
django-sendsms==0.3.1
djangorestframework==3.11.0
flake8==3.4.1
idna==2.9
isort==4.2.15
mccabe==0.6.1
phonenumbers==8.12.3
py==1.8.1
pycodestyle==2.3.1
pyflakes==1.5.0
pytest==2.9.1
pytest-cov==1.8.1
pytest-django==2.9.1
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
